+++ src/knowlift/web/static/js/utils.js
