+++ tests/test_web.py
@@ -29,7 +29,7 @@ from unittest import mock
 import flask
 
 # Project specific
-import knowlift
+from knowlift import web
 
 
 def check_membership(text, *strings):
@@ -44,7 +44,7 @@ class TestAppFactory(unittest.TestCase):
 
     def test_create_app_no_env(self):
         """Test creating the app without specifying an environment."""
-        app = knowlift.create_app()
+        app = web.create_app()
         self.assertIsInstance(app, flask.Flask)
         self.assertFalse(app.debug)
 
@@ -52,7 +52,7 @@ class TestAppFactory(unittest.TestCase):
         """Test creating the app with a specific environment."""
         for env in ['production', 'development', 'test']:
             with self.subTest(env=env):
-                app = knowlift.create_app(env=env)
+                app = web.create_app(env=env)
                 self.assertIsInstance(app, flask.Flask)
                 if env == 'development':
                     self.assertTrue(app.debug)
@@ -65,7 +65,7 @@ class TestDBConnection(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app context."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
         self.app_context = self.app.app_context()
         self.app_context.push()
@@ -78,25 +78,25 @@ class TestDBConnection(unittest.TestCase):
         """Test that getting a connection outside of an app context raises an error."""
         self.app_context.pop()
         with self.assertRaises(RuntimeError):
-            knowlift.get_connection()
+            web.get_connection()
         self.app_context.push()
 
     def test_get_and_close_connection(self):
         """Test getting and closing a database connection within an app context."""
-        with self.app.test_request_context():
+        with self.app.app_context():
             self.assertIsNone(getattr(flask.g, 'db', None))
 
-            conn = knowlift.get_connection()
+            conn = web.get_connection()
             self.assertIsNotNone(conn)
             self.assertIn('db', flask.g)
 
-            same_conn = knowlift.get_connection()
+            same_conn = web.get_connection()
             self.assertIs(conn, same_conn)
 
         self.assertIsNone(getattr(flask.g, 'db', None))
 
-    @mock.patch('knowlift.flask.current_app.logger')
-    @mock.patch('knowlift.views.number_distance.play')
+    @mock.patch('knowlift.web.flask.current_app.logger')
+    @mock.patch('knowlift.web.views.number_distance.play')
     def test_close_connection_on_exception(self, play_mock, logger_mock):
         """Test that the connection is closed when an exception occurs."""
         play_mock.side_effect = Exception("This is a test exception")
@@ -115,7 +115,7 @@ class TestIndexPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_index_page(self):
@@ -159,7 +159,7 @@ class TestAboutPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_about_page(self):
@@ -184,7 +184,7 @@ class TestCustomInternalErrorPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
         self.payload = {
             'left_glyph': '[',
@@ -225,7 +225,7 @@ class TestCustomNotFoundPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_custom_not_found_page(self):
@@ -255,7 +255,7 @@ class TestGradePage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_grade_page(self):
@@ -288,7 +288,7 @@ class TestLadderPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_ladder_page(self):
@@ -313,7 +313,7 @@ class TestLegalPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_legal_page(self):
@@ -361,7 +361,7 @@ class TestPlayPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
 
     def test_get_not_allowed(self):
@@ -428,7 +428,7 @@ class TestResultPage(unittest.TestCase):
 
     def setUp(self):
         """Set up a test app and client."""
-        self.app = knowlift.create_app(env='test')
+        self.app = web.create_app(env='test')
         self.client = self.app.test_client()
         self.post_data = {
             'left_glyph': '[',
